Fixed issue #82 (Initial delta of Time.fps on Mac)
The issue can be found here: https://github.com/switchface/helm/issues/82

The problem was that every function that called every'' (inside
FRP.Helm.Time) would get back the current timestamp as a first delta.

It's much more reasonable to have it return `0` as a first delta,
instead; especially in games where an initial update of 1419151612000
milliseconds could screw up the initial state. As an example, assuming a
simple snake game, an initial update of 1419151612 seconds in a snake
game could move the snake out of the bounds (and therefore make the
player lose) at the first "tick".

This commit aims at solving exactly this problem, but gathering the
initial timestamp to pass as the initial argument of `transfer`.